+++ at_secondary/at_secondary_server/lib/src/notification/notification_manager_impl.dart
@@ -59,4 +59,9 @@ class NotificationManager implements NotificationManagerSpec {
   bool isNotificationAccepted() {
     return true;
   }
+
+  @override
+  Future<void> remove(String notificationId) async {
+    await AtNotificationKeystore.getInstance().remove(notificationId);
+  }
 }