+++ at_server_spec/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 3.0.8
+- Exported 'NotifyRemove' verbs in package export
 ## 3.0.7
 - Updated 'info' verb documentation to cover new functionality
 ## 3.0.6